исправлены косяки
В тестах следует использовать человекочитаемые входные данные, а не беспорядочную последовательность символов
Это касается как параметров командной строки, так и содержимого входных файлов


Позиция в контейнерах string, vector и других, а также размер имеют тип size_t
это не то же самое, что int.
size_t на 64-битных платформах 64-битный, а int - 32-битный.
Твои приложения не смогут корректно работать в 64 битной конфигурации

+++ lab1/replace/replace.cpp
@@ -29,8 +29,8 @@ bool CheckFiles(ifstream &inputFile, ofstream &outputFile)
 
 void ReplaceStringInRow(string & rowFile, string const& searchString, string const& replaceString)
 {
-	int positionReplacedWords = rowFile.find(searchString);
-	int startSearchPosition = positionReplacedWords;
+	size_t positionReplacedWords = rowFile.find(searchString);
+	size_t startSearchPosition = positionReplacedWords;
 	int countMissingChar = 0;
 	while (positionReplacedWords != -1)
 	{
